Add a new "Initial file type" option in main config
Allows to select "Asm" or "Ascii" as the file type that will be
selected in the disk menu of the assembler when Compass boots.

The "rounded corners" option was stored in its own byte, now this
option is converted to a bit field that stores both the "rounded
corners" option and this new "initial file type" option.

Also:

- Fixed a bug that was always applying square corners
when returning from the shell.

- "Save configuration" text now changes to "Saving..." while saving.