Fix: memory search always performed on boot.
The bug was introduced when the configuration was changed to be
saved on a separate file. The availability for saved memory config
was checked before loading the settings, thus it was always
seeing that no config was available.

The fix consists of loading the configuration early on, before
performing the memory checks.