Use mappers table when available to generate ROMRAM table
If the mapper support routines are available, examine the mappers
table in order to fill the entries of the ROMRAM table for the
mapper slots at boot time, so that mapper sizes aren't calculated
again.

Also, remove the last direct usages of the mapper ports when the
mapper support routines are available.